Add files via upload
Added the AddGames Form to let users add new games to the database. Need to fix how the image is saved in order for it to be saved in the GameForms package with the .jpg images so user entries can be used in the FillForms class like the other records in the database.